% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Function to add two numeric values}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{numeric value 1}

\item{y}{numeric value 2, defaults to 10 if not specified}
}
\value{
summed result x + y
}
\description{
Function to add two numeric values
}
\examples{
my_add(1, 12)
my_add(10)
}
